relocated types and constants into new files
