+++ constants/index.ts
@@ -1,0 +1,1 @@
+export const userCount: number = 200;
