+++ pages/api/people.ts
@@ -1,18 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { faker } from "@faker-js/faker";
+import { User, Response } from "@/types";
+import { userCount } from "@/constants";
 
-export type User = {
-  name: string;
-  email: string;
-  title: string;
-  role: string;
-};
-
-type Response = {
-  users: User[],
-  numOfPages: number
-}
-export const userCount: number = 200;
 /**
  * TODO: Prepare an endpoint to return a list of users
  * User faker.js or similar library to generate fake data, the minimal number of users is 100
@@ -28,7 +18,7 @@ export const userCount: number = 200;
 
 const generateUsers = (count: number): User[] => {
   const userList: User[] = [];
-   
+
   for (let i = 0; i < count; i++) {
     const firstName: string = faker.person.firstName();
     const lastName: string = faker.person.lastName();
@@ -44,8 +34,8 @@ const generateUsers = (count: number): User[] => {
     };
     userList.push(newUser);
   }
-  
-  userList.sort((a,b) => a.name.localeCompare(b.name));
+
+  userList.sort((a, b) => a.name.localeCompare(b.name));
   return userList;
 };
 
@@ -55,11 +45,11 @@ export default function handler(
 ) {
   const { page = 1 } = req.query;
   const parsedPage = parseInt(page as string, 10);
-  
+
   const sortedUserList = generateUsers(userCount);
   const usersPerPage = 10;
   const numOfPages = Math.ceil(userCount / usersPerPage);
-  const startIndex = (parsedPage -1) * 10;
+  const startIndex = (parsedPage - 1) * 10;
   const paginatedUserList = sortedUserList.slice(startIndex, startIndex + 10);
-  res.status(200).json({users: paginatedUserList, numOfPages});
+  res.status(200).json({ users: paginatedUserList, numOfPages });
 }
