+++ types/index.ts
@@ -1,0 +1,11 @@
+export type User = {
+    name: string;
+    email: string;
+    title: string;
+    role: string;
+  };
+  
+export type Response = {
+    users: User[],
+    numOfPages: number
+}
